+++ crates/ruma-client-api/tests/uiaa.rs
@@ -127,7 +127,7 @@ fn deserialize_uiaa_info() {
     assert_eq!(info.session.as_deref(), Some("xxxxxx"));
     let auth_error = info.auth_error.unwrap();
     assert_matches!(auth_error.kind, ErrorKind::Forbidden { .. });
-    assert_eq!(auth_error.message, "Invalid password");
+    assert_eq!(auth_error.message.as_deref(), Some("Invalid password"));
     assert_eq!(
         from_json_str::<JsonValue>(info.params.unwrap().get()).unwrap(),
         json!({
@@ -210,7 +210,7 @@ fn try_uiaa_response_from_http_response() {
     assert_eq!(info.session.as_deref(), Some("xxxxxx"));
     let auth_error = info.auth_error.unwrap();
     assert_matches!(auth_error.kind, ErrorKind::Forbidden { .. });
-    assert_eq!(auth_error.message, "Invalid password");
+    assert_eq!(auth_error.message.as_deref(), Some("Invalid password"));
     assert_eq!(
         from_json_str::<JsonValue>(info.params.unwrap().get()).unwrap(),
         json!({